Fix compilation errors and clippy warnings
- Fix type mismatch in entropy_analysis.rs (u32 vs usize comparison)
- Replace or_insert_with(Vec::new) with or_default() throughout codebase
- Remove unnecessary type casts
- Remove needless borrows in function calls
- Add Default trait implementations for CodeQualityAnalyzer, DependencyAnalyzer,
  McpTransportServer, StringTracker, and VulnerabilityDetector
- Fix manual clamp pattern with clamp() method
- Fix manual strip pattern with strip_suffix()
- Simplify collapsible match pattern in disassembly.rs
- Fix enum naming conventions (MIT->Mit, GPL->Gpl, LGPL->Lgpl, BSD->Bsd)
- Add clippy allow attribute for false positive recursion warning

+++ src/binary_parser.rs
@@ -185,7 +185,7 @@ fn parse_mach(mach: goblin::mach::Mach) -> Result<BinaryInfo> {
     match mach {
         Mach::Binary(mach_o) => {
             info.architecture = if mach_o.is_64 { "x86_64" } else { "x86" }.to_string();
-            info.entry_point = Some(mach_o.entry as u64);
+            info.entry_point = Some(mach_o.entry);
             for segment in &mach_o.segments {
                 for (section, _) in &segment.sections()? {
                     info.sections.push(SectionInfo {
